requirements.txt
aiogram==2.12.1
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
ipinfo==4.1.0
lml==0.1.0
loguru==0.5.3
multidict==5.1.0
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pymemcache==3.4.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
texttable==1.6.3
typing-extensions==3.7.4.3
urllib3==1.26.4
xlrd==1.2.0
xlwt==1.3.0
yarl==1.6.3
